requirements.txt
# aiobotocore==2.4.2
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
#awscli==1.27.85
bcrypt==4.0.1
beautifulsoup4==4.11.2
boto3
botocore==1.29.85
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
cramjam==2.6.2
cryptography==39.0.1
cx-Oracle==8.3.0
docutils==0.16
fastparquet==2023.2.0
frozenlist==1.3.3
fsspec==2023.1.0
greenlet==2.0.2
iaccess==0.4.2
idna==3.4
jmespath==1.0.1
lxml==4.9.2
minio==7.1.14
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.1
packaging==23.0
pandas==1.5.3
paramiko==3.0.0
psycopg2==2.9.3
pyaml==21.10.1
pyarrow==11.0.0
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pyodbc==4.0.35
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
redshift-connector==2.0.910
requests==2.28.2
rsa==4.7.2
scramp==1.4.4
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlalchemy-redshift==0.8.12
urllib3==1.26.14
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.8.2
